Move all the watching to the elementor editor integration. Initialize in elementor.js
